change short to byte type on TransformJ

+++ acc/Tokens/TransformPushJ.cs
@@ -5,9 +5,9 @@
 
     public class TransformPushJ : TransformationContext
     {
-        public TransformPushJ(string value, short cellDev, short ActionDev)
+        public TransformPushJ(string value, byte cellDev, byte ActionDev)
         {
-            Instructions = value.Select(x => new push_a(cellDev, ActionDev, (short) x)).Cast<Instruction>().ToArray();
+            Instructions = value.Select(x => new push_a(cellDev, ActionDev, x)).Cast<Instruction>().ToArray();
         }
     }
 }
